Add test canvas version of a zombie
doing this allows for the posibility of being able to easily generate random zombies with random looks etc. furthermore, you are able to pause the animation - which is a lot easier than using a gif. In the future, all the zombie parts could be loaded onto a single sprite - or even use svg and store that in a json file?